move simapp/helpers to test_helpers
