--- simapp/helpers/test_helpers.go
@@ -1,77 +1,0 @@
-package helpers
-
-import (
-	"math/rand"
-	"time"
-
-	"github.com/tendermint/tendermint/crypto"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/simulation"
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-	authsign "github.com/cosmos/cosmos-sdk/x/auth/signing"
-)
-
-// SimAppChainID hardcoded chainID for simulation
-const (
-	DefaultGenTxGas = 1000000
-	SimAppChainID   = "simulation-app"
-)
-
-// GenTx generates a signed mock transaction.
-func GenTx(gen client.TxGenerator, msgs []sdk.Msg, feeAmt sdk.Coins, gas uint64, chainID string, accnums []uint64, seq []uint64, priv ...crypto.PrivKey) (sdk.Tx, error) {
-	sigs := make([]signing.SignatureV2, len(priv))
-
-	// create a random length memo
-	r := rand.New(rand.NewSource(time.Now().UnixNano()))
-
-	memo := simulation.RandStringOfLength(r, simulation.RandIntBetween(r, 0, 100))
-
-	signMode := gen.SignModeHandler().DefaultMode()
-
-	for i, p := range priv {
-		sigs[i] = signing.SignatureV2{
-			PubKey: p.PubKey(),
-			Data: &signing.SingleSignatureData{
-				SignMode: signMode,
-			},
-		}
-	}
-
-	tx := gen.NewTxBuilder()
-	err := tx.SetMsgs(msgs...)
-	if err != nil {
-		return nil, err
-	}
-	err = tx.SetSignatures(sigs...)
-	if err != nil {
-		return nil, err
-	}
-	tx.SetMemo(memo)
-	tx.SetFeeAmount(feeAmt)
-	tx.SetGasLimit(gas)
-	for i, p := range priv {
-		// use a empty chainID for ease of testing
-		signerData := authsign.SignerData{
-			ChainID:         chainID,
-			AccountNumber:   accnums[i],
-			AccountSequence: seq[i],
-		}
-		signBytes, err := gen.SignModeHandler().GetSignBytes(signMode, signerData, tx.GetTx())
-		if err != nil {
-			panic(err)
-		}
-		sig, err := p.Sign(signBytes)
-		if err != nil {
-			panic(err)
-		}
-		sigs[i].Data.(*signing.SingleSignatureData).Signature = sig
-		err = tx.SetSignatures(sigs...)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	return tx.GetTx(), nil
-}
